norma fix

+++ srcs/md_read.c
@@ -6,11 +6,11 @@
 /*   By: uhand <uhand@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/20 17:20:54 by uhand             #+#    #+#             */
-/*   Updated: 2020/11/29 22:20:32 by uhand            ###   ########.fr       */
+/*   Updated: 2020/11/30 12:06:10 by uhand            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "../includes/md5.h"
+#include "../includes/md5.h"
 
 static int	read_string(t_read *rd, t_common alg)
 {
@@ -18,17 +18,17 @@ static int	read_string(t_read *rd, t_common alg)
 	char	*block_ptr;
 
 	i = 0;
-	if	(rd->str_msg[0] == '\0')
+	if (rd->str_msg[0] == '\0')
 		return (uppend_block(rd, i, alg));
 	block_ptr = (char*)rd->x;
-	while (rd->str_msg[0] != '\0' && i < alg.block_size)
+	while (rd->str_msg[0] != '\0' && i < (size_t)alg.block_size)
 	{
 		block_ptr[i] = rd->str_msg[0];
 		rd->read_count++;
 		rd->str_msg++;
 		i++;
 	}
-	if (i < alg.block_size)
+	if (i < (size_t)alg.block_size)
 		return (uppend_block(rd, i, alg));
 	return (1);
 }
@@ -39,7 +39,7 @@ static int	copy_block(t_read *rd, t_common alg)
 
 	cpy_size = (rd->ret < alg.block_size) ? rd->ret : alg.block_size;
 	ft_memcpy(rd->x, rd->ptr, cpy_size);
-	if (rd->i + cpy_size >= alg.read_size)
+	if (rd->i + cpy_size >= (int)alg.read_size)
 	{
 		rd->i = 0;
 		rd->ptr = NULL;
@@ -61,19 +61,19 @@ static int	read_from_fd(t_read *rd, t_common alg)
 	if (!rd->ptr && !(rd->ret = read(rd->prc->fd, &rd->buf, alg.read_size)))
 		return (uppend_block(rd, rd->ret, alg));
 	else if (rd->ptr)
-		return(copy_block(rd, alg));
+		return (copy_block(rd, alg));
 	else if (rd->ret && rd->ret != -1)
 	{
 		if (rd->print)
 			write(1, rd->buf, rd->ret);
 		rd->ptr = rd->buf;
-		return(copy_block(rd, alg));
+		return (copy_block(rd, alg));
 	}
 	else if (rd->ret && rd->ret == -1)
 	{
 		ft_printf("ft_ssl: %s: %s: %s\n", rd->cmd_name, rd->prc->file_name, \
 			strerror(errno));
-		exit (-1);
+		exit(-1);
 	}
 	return (0);
 }
